changes urls of pages starting with /static to / and made it backwards compactible
