+++ src/pages/terms.tsx
@@ -189,8 +189,8 @@ const Terms = () => (
         <Box>
           <RelatedTopics
             topics={[
-              { name: 'Privacy Policy', url: '/static/privacy' },
-              { name: 'Guidelines', url: '/static/guidelines' },
+              { name: 'Privacy Policy', url: '/privacy' },
+              { name: 'Guidelines', url: '/guidelines' },
             ]}
           />
         </Box>
